+++ src/icalendar/parser.py
@@ -378,7 +378,7 @@ class Contentlines(list):
         return b'\r\n'.join(line.to_ical() for line in self if line) + b'\r\n'
 
     @classmethod
-    def from_ical(cls, st):
+    def from_ical(cls, st, strict=False):
         """Parses a string into content lines.
         """
         st = to_unicode(st)