+++ CHANGES.rst
@@ -20,6 +20,12 @@ Bug fixes:
 - Fix handling of VTIMEZONEs with subcomponents with the same DTSTARTs.
   [geier]
 
+- Make every from_ical() method accept an optional `strict=False` parameter
+  [geier]
+
+- Don't break on UTCOffsets > 24 hours, even though they make no sense, at least
+  one major calendar service provider uses those [geier]
+
 - *add item here*