+++ CHANGES.rst
@@ -1,6 +1,22 @@
 Changelog
 =========
 
+3.11.5 (unreleased)
+-------------------
+
+Breaking changes:
+
+- *add item here*
+
+New features:
+
+- *add item here*
+
+Bug fixes:
+
+- *add item here*
+
+
 3.11.4 (2017-05-10)
 -------------------